requirements.txt
django
django-phonenumber-field
